Avoid circular import, removed load_test_dat command and related test
